requirements.txt
asgiref==3.6.0
attrs==23.1.0
Django==4.2.1
djangorestframework==3.14.0
drf-nested-routers==0.93.4
drf-spectacular==0.26.2
env==0.1.0
inflection==0.5.1
jsonschema==4.17.3
pyrsistent==0.19.3
pytz==2023.3
PyYAML==6.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
